refactor: eliminate DRY violations across server services and client hooks
- Extract getPaginationParams + pickAllowed into server/utils/queryHelpers.js,
  replacing 9 inline pagination patterns and 6 inline field-filter patterns
- Extract shared useFormData hook; useCarForm and useServiceForm (identical files)
  now re-export from it
- Move serviceStatusOptions to client/src/utils/serviceConstants.js, replacing
  duplicate definitions in carValidation.js and serviceValidation.js
