+++ client/src/pages/servicesAdmin/utils/serviceValidation.js
@@ -12,11 +12,4 @@ export const serviceFilterFn = (item, value) =>
   item.paid.toString().includes(value) ||
   item.status.includes(value);
 
-/**
- * Service status options
- */
-export const serviceStatusOptions = [
-  { value: "pending", label: "Pending" },
-  { value: "done", label: "Done" },
-  { value: "on-work", label: "On work" },
-];
+export { serviceStatusOptions } from "../../../utils/serviceConstants";
