+++ client/src/utils/serviceConstants.js
@@ -1,0 +1,5 @@
+export const serviceStatusOptions = [
+  { value: "pending", label: "Pending" },
+  { value: "done", label: "Done" },
+  { value: "on-work", label: "On work" },
+];
