+++ server/services/appointmentService.js
@@ -1,6 +1,7 @@
 import Appointment from "../models/Appointment.js";
 import { templatePhone } from "../utils/templates.js";
 import { createError } from "../utils/error.js";
+import { pickAllowed } from "../utils/queryHelpers.js";
 
 const ALLOWED_APPOINTMENT_CREATE_FIELDS = ['clientName', 'email', 'phone', 'date', 'time', 'notes', 'user'];
 
@@ -8,9 +9,7 @@ const createAppointment = async (req) => {
   const { phone } = req.body;
   const formattedPhone = phone ? templatePhone(phone) : phone;
 
-  const safeBody = Object.fromEntries(
-    Object.entries(req.body).filter(([k]) => ALLOWED_APPOINTMENT_CREATE_FIELDS.includes(k))
-  );
+  const safeBody = pickAllowed(req.body, ALLOWED_APPOINTMENT_CREATE_FIELDS);
 
   const newAppointment = new Appointment({ ...safeBody, phone: formattedPhone });
   const savedAppointment = await newAppointment.save();
