+++ server/services/carService.js
@@ -2,6 +2,7 @@ import Car from "../models/Car.js";
 import User from "../models/User.js";
 import { templateCar } from "../utils/templates.js";
 import { createError } from "../utils/error.js";
+import { getPaginationParams, pickAllowed } from "../utils/queryHelpers.js";
 
 const ALLOWED_CAR_POPULATE_FIELDS = ['services', 'owner'];
 const ALLOWED_CAR_UPDATE_FIELDS = ['numberPlate', 'km', 'brand'];
@@ -23,9 +24,7 @@ const createCar = async (req) => {
 };
 
 const updateCar = async (req) => {
-  const safeBody = Object.fromEntries(
-    Object.entries(req.body).filter(([key]) => ALLOWED_CAR_UPDATE_FIELDS.includes(key))
-  );
+  const safeBody = pickAllowed(req.body, ALLOWED_CAR_UPDATE_FIELDS);
   if (safeBody.numberPlate) {
     safeBody.numberPlate = templateCar(safeBody.numberPlate);
   }
@@ -57,8 +56,7 @@ const getCar = async (req) => {
 };
 
 const getCars = async (req) => {
-  const limit = Math.min(parseInt(req.query.limit) || 500, 500);
-  const page = Math.max(parseInt(req.query.page) || 1, 1);
+  const { limit, page } = getPaginationParams(req);
   const cars = await Car.find().populate("owner").skip((page - 1) * limit).limit(limit);
   return cars;
 };
@@ -68,15 +66,13 @@ const getCarsByType = async (req) => {
   if (!ALLOWED_CAR_POPULATE_FIELDS.includes(type)) {
     throw createError(400, `Invalid populate field. Allowed: ${ALLOWED_CAR_POPULATE_FIELDS.join(', ')}`);
   }
-  const limit = Math.min(parseInt(req.query.limit) || 500, 500);
-  const page = Math.max(parseInt(req.query.page) || 1, 1);
+  const { limit, page } = getPaginationParams(req);
   const cars = await Car.find().populate(type).skip((page - 1) * limit).limit(limit);
   return cars;
 };
 
 const getCarsWithService = async (req) => {
-  const limit = Math.min(parseInt(req.query.limit) || 500, 500);
-  const page = Math.max(parseInt(req.query.page) || 1, 1);
+  const { limit, page } = getPaginationParams(req);
   const cars = await Car.find().populate("services").skip((page - 1) * limit).limit(limit);
   return cars;
 };
