+++ server/services/reviewService.js
@@ -1,19 +1,17 @@
 import Review from "../models/Review.js";
+import { getPaginationParams, pickAllowed } from "../utils/queryHelpers.js";
 
 const ALLOWED_REVIEW_FIELDS = ['name', 'description', 'stars'];
 
 const createReview = async (req) => {
-  const safeBody = Object.fromEntries(
-    Object.entries(req.body).filter(([k]) => ALLOWED_REVIEW_FIELDS.includes(k))
-  );
+  const safeBody = pickAllowed(req.body, ALLOWED_REVIEW_FIELDS);
   const newReview = new Review(safeBody);
   const savedReview = await newReview.save();
   return savedReview;
 };
 
 const getReviews = async (req) => {
-  const limit = Math.min(parseInt(req.query.limit) || 500, 500);
-  const page = Math.max(parseInt(req.query.page) || 1, 1);
+  const { limit, page } = getPaginationParams(req);
   const reviews = await Review.find().skip((page - 1) * limit).limit(limit);
   return reviews;
 };
